test(use-idle): recategorize and clean up browser tests
